Unified naming style, moved S7 collections to a separate file
