--- s7/factories/__init__.py
@@ -1,7 +1,0 @@
-"""Public factory functions for creating SOFT7-related classes and instances."""
-
-from __future__ import annotations
-
-from .entity_factory import create_entity
-
-__all__ = ("create_entity",)
